+++ client/src/views/tabs/favorites.js
@@ -5,9 +5,9 @@ import { StyleSheet } from 'react-native';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import favoriteActions from '../actions/favorite-actions';
-import List from '../components/list/List.component';
-import colors from '../components/style/colors';
+import favoriteActions from '../../actions/favorite-actions';
+import List from '../../components/list/List.component';
+import colors from '../../components/style/colors';
 
 class Favorites extends React.Component {
   constructor(props) {
